Failed attempt at fixing the scheduler.
The problem was that the calls were sometimes done with a delay of 32
ms. instead of 16.

Using the update per frame from Axmol and tracking time with the delta
time argument it passes to the update function did not solve the
problem.

Tracking the time by hand was of no help.

Applying a fixed delta time of 1/60 seconds in each tick did not help
either.

Maybe the re-scheduling part from iscool::schedule is wrong?

Anyway, there is no way to be precise below 1/60 seconds. Depending on
the precision of the clock and the current load, a call scheduled in
1/60 seconds may happen in the next tick or the following. The next
tick may be a bit short, thus pushing the call in the following tick,
then the following tick may have some work to do before doing the
call, thus delaying it further.