Explicit game over message from the server to the clients.
The clients used to quit the game as soon as they reached the game
over locally, preventing further synchronization with the
server. Specifically, the server would not receive confirmation
that the client is synchronized with the game over state.

Now the clients keep going until they receive an explicit game over
message from the server.