Fix a case of a falling block not being solid.
When the block was sealed, a static wall was added in the arena, which
also set the is_solid flag for the cell. Then the entity from the
cell, here a flame, was removed by the falling block. This caused the
is_solid flag to be unset. Now we distinguish solid entities and
static walls.